remove instance meta data when unloading patcher instance
